Moved *OrCatch request functions in their own file
